b2b_one/requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
cryptography==39.0.0
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-datatables-view==1.20.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-invitations==2.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
drf-spectacular==0.25.1
graphviz==0.20.1
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
phonenumbers==8.13.4
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
twilio==7.16.2
uritemplate==4.1.1
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
